+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - change settings amount popup
 - spends storage
+- add-spends popup
 
 ## [3.1.0] - 2021-12-24
 ### Added